Renamed module to barcode
