+++ aamva/barcode/utils.py
