+++ tests/barcode/test_subfile.py
@@ -1,6 +1,6 @@
 import pytest
 
-import aamva.barcode_file.subfile as subfile
+import aamva.barcode.subfile as subfile
 
 
 testdata_ids = ("1 element", "multi element", "multi element with empty")
