+++ tests/barcode/test_utils.py
@@ -1,6 +1,6 @@
 import pytest
 
-import aamva.barcode_file.utils as utils
+import aamva.barcode.utils as utils
 
 
 def test_can_trim_before():
